requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
requests==2.7.0
wsgiref==0.1.2
-e git+https://github.com/os/slacker#egg=slacker
Pillow==3.2.0
psycopg2==2.6.1
Flask-Heroku-Cacheify>=1.3
pylibmc>=1.2.3
FontTools==3.0
